backend: move remap into own module
