+++ backend/fixtures/useFixtures.js
@@ -4,7 +4,7 @@ import { getCollection } from '../api/utils/getCollection'
 import { createLog } from '../infrastructure/log/createLog'
 import { SyncState } from '../contexts/sync/SyncState'
 import { Meteor } from 'meteor/meteor'
-import { runRemap } from '../server/reamp'
+import { runRemap } from '../server/remap'
 
 const fixturesIsActive = Meteor.settings.useFixtures
 const debug = createLog({ name: 'useFixtures', type: 'debug' })
